Moved the code in __init__.py to a module in order to fix the AppRegistryNotReady error in Django 1.9.
